Attribute Macro obce_macro::mock

#[mock]

Chain extension mocking utility.

Description

You can use #[obce::mock] to automatically generate register_chain_extensions function, which accepts a context and automatically registers mocked chain extension methods for off-chain ink! smart contract testing.

// ink! smart contract definition is omitted.

#[obce::definition]
pub trait MyChainExtension {
    fn test_method(&mut self, val: u32, another_val: u32) -> u32;
}

#[obce::mock]
impl MyChainExtension for () {
    fn test_method(&mut self, val: u32, another_val: u32) -> u32 {
        val + another_val
    }
}

#[test]
fn call_contract() {
    register_chain_extensions(());
    let mut contract = SimpleContract::new();
    assert_eq!(contract.call_test_method(100, 200), 300);
}

When using #[obce::mock], you are not required to fill every single method for testing. Glue code to register chain extension methods will only apply to those methods, that you listed in a mock macro call:

#[obce::definition]
pub trait MyChainExtension {
    fn first_method(&mut self, val: u32) -> u32;
    fn second_method(&mut self) -> u64;
}

#[obce::mock]
impl MyChainExtension for () {
    fn first_method(&mut self, val: u32) -> u32 {
        // ...
    }

    // second_method is not required to be present here
}

Context

The item that you implement your definition trait for becomes your testing context.

You will receive the same testing context when calling methods multiple times, thus it can be used as your chain extension testing state.

General guidelines

Since #[obce::mock] is designed for off-chain testing, you are limited by off-chain testing facilities that ink! library provides.

Complete example

#[obce::definition(id = 123)]
pub trait ChainExtension {
    fn method(&mut self, val: u32, another_val: u32) -> u32;

    #[obce(id = 456)]
    fn another_method(&mut self, val: u32) -> u32;
}

struct MyChainExtension;

impl ChainExtension for MyChainExtension {}

#[ink::contract]
mod simple_contract {
    use crate::{
        ChainExtension,
        MyChainExtension,
    };

    #[ink(storage)]
    pub struct SimpleContract {}

    impl SimpleContract {
        #[ink(constructor)]
        pub fn new() -> Self {
            SimpleContract {}
        }

        #[ink(message)]
        pub fn call_method(&mut self, val: u32, another_val: u32) -> u32 {
            MyChainExtension.method(val, another_val)
        }

        #[ink(message)]
        pub fn call_another_method(&mut self, val: u32) -> u32 {
            MyChainExtension.another_method(val)
        }
    }
}

mod simple_test {
    #[obce::mock]
    impl crate::ChainExtension for () {
        fn method(&mut self, val: u32, another_val: u32) -> u32 {
            val + another_val
        }
    }

    #[test]
    fn call_contract() {
        register_chain_extensions(());
        let mut contract = crate::simple_contract::SimpleContract::new();
        assert_eq!(contract.call_method(100, 200), 300);
    }
}